Fix UI tests breaking with embedded EMS
The solution was to copy the Ticket SDK framework in ExampleApp using a
Build Phase.

Also tidies up so that the ExampleApp uses the Ticket SDK like a
framework now. I had to mark a tonne of functions and variables public
to make them accessible outside of the framework. This is something we
need to continue doing.