Color to reflect last price
reflecting if the price is up or down based on the day changed percentage.

There is a lag for the color reflection I can't seem to fix, and there is the issue of cost with the server constantly retrieving the day change value, so I am unsure on whether or not this should be a feature for the last price.